Refactored flag code into a more reusable 'Sprite' class.
